+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Debug` derive for `FromClient`.
 
+### Fixed
+
+- Replication of entities without components before connection.
+
 ## [0.35.1] - 2025-10-03
 
 ### Added